refactor: Изменено расположение enum в пакет с моделями, проверка юзера осуществляется теперь в сервисах
