+++ src/main/java/ru/yandex/practicum/filmorate/model/feedTypes/Operation.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.storage.feed;
+package ru.yandex.practicum.filmorate.model.feedTypes;
 
 public enum Operation {
     REMOVE,
