+++ src/main/java/ru/yandex/practicum/filmorate/storage/feed/UserFeedDbStorage.java
@@ -2,11 +2,11 @@ package ru.yandex.practicum.filmorate.storage.feed;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Component;
-import ru.yandex.practicum.filmorate.Exceptions.NotFoundException;
 import ru.yandex.practicum.filmorate.model.Feed;
+import ru.yandex.practicum.filmorate.model.feedTypes.Event;
+import ru.yandex.practicum.filmorate.model.feedTypes.Operation;
 import ru.yandex.practicum.filmorate.storage.user.UserStorage;
 
 import java.time.Instant;
@@ -36,21 +36,14 @@ public class UserFeedDbStorage {
 
     public List<Feed> getFeedByUserId(Long userId) {
         String sqlQuery = "SELECT * FROM feeds WHERE user_id = ?";
-
-        if (userStorage.getUserById(userId) != null) {
-            log.info("Действе юзера с id{}", userId);
-
-            return jdbcTemplate.query(sqlQuery, new Object[]{userId}, (rs, rowNum) -> new Feed(
-                            rs.getLong("feed_timestamp"),
-                            rs.getLong("user_id"),
-                            Event.valueOf(rs.getString("event_type")),
-                            Operation.valueOf(rs.getString("operation")),
-                            rs.getLong("event_id"),
-                            rs.getLong("entity_id")
-                    )
-            );
-        } else {
-            throw  new NotFoundException(HttpStatus.NOT_FOUND.toString());
-        }
+        return jdbcTemplate.query(sqlQuery, new Object[]{userId}, (rs, rowNum) -> new Feed(
+                        rs.getLong("feed_timestamp"),
+                        rs.getLong("user_id"),
+                        Event.valueOf(rs.getString("event_type")),
+                        Operation.valueOf(rs.getString("operation")),
+                        rs.getLong("event_id"),
+                        rs.getLong("entity_id")
+                )
+        );
     }
 }
