+++ src/main/resources/schema.sql
@@ -1,3 +1,18 @@
+DROP TABLE IF EXISTS `film_likes` CASCADE;
+DROP TABLE IF EXISTS `ratings_mpa` CASCADE;
+DROP TABLE IF EXISTS `directors` CASCADE;
+DROP TABLE IF EXISTS `film_directors` CASCADE;
+DROP TABLE IF EXISTS `genres` CASCADE ;
+DROP TABLE IF EXISTS `friends` CASCADE;
+DROP TABLE IF EXISTS `films` CASCADE;
+DROP TABLE IF EXISTS `users` CASCADE;
+DROP TABLE IF EXISTS `film_genres` CASCADE;
+DROP TABLE IF EXISTS `reviews` CASCADE;
+DROP TABLE IF EXISTS `review_reactions` CASCADE;
+DROP INDEX IF EXISTS review_reactions_user_id_index;
+DROP INDEX IF EXISTS review_reactions_review_id_index;
+DROP TABLE IF EXISTS feeds CASCADE;
+
 CREATE TABLE IF NOT EXISTS ratings_mpa
 (
     id          int generated by default as identity primary key,
